code/requirements.txt
# 核心爬虫框架
scrapy>=2.11.0
scrapy-redis>=0.7.0
scrapy-splash>=0.3.0

# 数据库
mysql-connector-python>=8.0.32
sqlalchemy>=2.0.0
alembic>=1.12.0
pymongo>=4.5.0

# 数据处理
beautifulsoup4>=4.12.0
lxml>=4.9.0
nltk>=3.8.0
spacy>=3.6.0
pandas>=2.0.0
numpy>=1.24.0

# 文本处理
regex>=2023.0.0
jieba>=0.42.1
wordcloud>=1.9.2

# HTTP和异步
requests>=2.31.0
aiohttp>=3.8.0
httpx>=0.24.0

# 日志和监控
loguru>=0.7.0
prometheus-client>=0.16.0

# Web框架
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.0.0

# 测试
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# 工具库
python-dotenv>=1.0.0
pytz>=2023.3
pydantic>=2.3.0
click>=8.1.0
colorama>=0.4.6
